+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add online users metric to admin app
 - improve performance of admin app
 - add changelog page to admin app
+- show 14 days as 2 weeks in how_long function
 
 ## [6.16.1] - 2021-09-02
 - fix bug in get_user()