+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- add additional chromedriver options to stabilize automated testing
 - add stop impersonation support for sessions with user override
 - add error log entry for low level errors
 - allow packages to be included before an app has been loaded