+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add error log entry for low level errors
 - allow packages to be included before an app has been loaded
 - add support for bootstrap icons
+- add app.theme_path member
 
 ## [6.17.0] - 2022-01-01
 - autofocus on username on login