+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.18.0] - 2022-03-02
 - add additional chromedriver options to stabilize automated testing
 - add stop impersonation support for sessions with user override
 - add error log entry for low level errors